requirements.txt
flask==3.0.3
requests==2.31.0
python-dotenv==1.0.0
openai==0.28.0
gunicorn==22.0.0
PyPDF2==3.0.1
Pillow==10.2.0
redis==4.6.0
jsonschema==3.2.0
tiktoken==0.3.3
sentence-transformers==2.2.2
numpy>=1.26.0
flask-cors==4.0.0
json5==0.9.6
pymongo==4.3.3  # Add this line